refactor: servlet->controller 변경 및 리팩터링
