+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -7,12 +7,12 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.catalina.session.Session;
 import org.apache.coyote.http11.HttpHeaders;
-import org.apache.coyote.http11.HttpServlet;
+import org.apache.coyote.http11.controller.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.StatusCode;
 
-public class LoginController extends HttpServlet {
+public class LoginController extends AbstractController {
 
     private static final String PREFIX = "static";
     private static final String SUFFIX = ".html";
